% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extrai_atividade}
\alias{extrai_atividade}
\title{Extrai uma atividade}
\usage{
extrai_atividade(
  project = "elsa",
  url = "https://api-otus.elsa.ufrgs.br/enterprises/data-extraction/activity",
  acronym,
  version,
  token,
  ...
)
}
\arguments{
\item{project}{O nome do projeto no Otus.}

\item{url}{O nome da url de extração de dados para a atividade desejada.
A url default utilizada é
"https://api-otus.elsa.ufrgs.br/enterprises/data-extraction/activity"}

\item{acronym}{O acrônimo da atividade para a qual se deseja extrair os dados.}

\item{version}{O número da versão da atividade para a qual se deseja extrair os dados.}

\item{token}{character o token do usuário.}
}
\description{
A função extrai_atividade extrai uma versão de um atividade
  através do api do Otus ELSA, retornando um data frame/tibble.
}
