% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extrai_atividade_lista}
\alias{extrai_atividade_lista}
\title{Extrai as versões de uma atividade}
\usage{
extrai_atividade_lista(
  project = "elsa",
  url = "https://api-otus.elsa.ufrgs.br/enterprises/data-extraction/activity",
  acronym,
  token,
  ...
)
}
\arguments{
\item{project}{O nome do projeto no Otus.}

\item{url}{O nome da url de extração de dados para a atividade desejada.
A url default utilizada é
"https://api-otus.elsa.ufrgs.br/enterprises/data-extraction/activity"}

\item{acronym}{O acrônimo da atividade para a qual se deseja extrair os dados.}

\item{token}{character o token do usuário.}
}
\description{
A função extrai_atividade_lista extrai todas as versões
  correspondentes de uma atividade
  através do api do Otus ELSA, retornando uma lista de data frames.
}
