% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extrai_versao}
\alias{extrai_versao}
\title{Extrai número de versões}
\usage{
extrai_versao(
  project = "elsa",
  url = "https://api-otus.elsa.ufrgs.br/enterprises/data-extraction/activity",
  acronym,
  token,
  ...
)
}
\arguments{
\item{project}{O nome do projeto no Otus.}

\item{url}{O nome da url de extração de dados para a atividade desejada.
A url default utilizada é
"https://api-otus.elsa.ufrgs.br/enterprises/data-extraction/activity"}

\item{acronym}{O acrônimo da atividade para a qual se deseja extrair os dados.}

\item{token}{character o token do usuário.}
}
\description{
A função extrai_versao extrai o número de versões (número total)
  e os números correspondentes das versões de uma atividade
  através do api do Otus ELSA, retornando uma lista.
}
